created utils file
