+++ webapp/maps_heatmaps/src/index.tsx
@@ -31,11 +31,12 @@ import ReactDOM from "react-dom";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import SidePanel from "./components/sidepanel";
+import { eraseAllMarkers, addMarkerWithListener } from "./clickInfoWindow";
 import {
-  eraseAllMarkers,
+  convertLatLngToGeopoint,
   convertGeopointToLatLon,
-  addMarkerWithListener,
-} from "./clickInfoWindow";
+  getRadius,
+} from "./utils";
 import { DateTime } from "./interface";
 
 let map: google.maps.Map, heatmap: google.maps.visualization.HeatmapLayer;
@@ -99,14 +100,7 @@ async function mapChanged() {
   const newCenter = new firebase.firestore.GeoPoint(lat, lng);
   const bounds = map.getBounds(); //map's current bounderies
   //TODO: check what should be the default radius value.
-  let newRadius = 2;
-  if (bounds) {
-    const meterRadius = google.maps.geometry.spherical.computeDistanceBetween(
-      bounds.getCenter(),
-      bounds.getNorthEast()
-    );
-    newRadius = meterRadius * 0.000621371192; //convert to miles
-  }
+  const newRadius = getRadius(bounds);
   if (timeOfLastRequest === timeOfRequest) {
     const queriedCollection = queryDB.getQueriedCollection(
       newCenter,
